Add expert logic for Your Contract has Expired
Cherry hover to the post-fight cutscene from the main subcon area,
skipping the fight:
https://github.com/CookieCat45/Archipelago-ahit/blob/0b41527c2278cdd1d29a6ef913ef2a5ae03533c7/worlds/ahit/Rules.py#L586

The logic is more difficult to represent due to YCHE now being
accessible from just the main subcon area without anything else in
expert difficulty. Because of this, YCHE has been moved to its own
location.

Toilet of Doom is now its own location due to being the only remaining
section. This allows for its logic to be simplified across the location
and section access_rules instead of having all the rules in the section.