Speed up receiving items/locations from an in-progress slot
When connecting to an in-progress slot, the onLocation and onItem
handlers have to receive all current items and cleared locations. This
happens separately from onClear, so re-enabling tracker updates at the
end of `onClearHandler()` would happen before the items and locations
are received.

Tracker updates are now disabled until 10ms after onClear is handled, by
running a 'busy sleep' asynchronously, and re-enabled on the first frame
after the 10ms duration is up.

This should give enough time for all items and locations to be received
before tracker updates are re-enabled.

The 10ms duration may want to be increased if it's not enough for slower
computers or if more items/locations are added to AHiT in the future.