Add Relics to the tracker
Relics are implemented as progressive items that add a new relic piece
to the icon for each relic of that type collected. The order the relics
are added to the icon is fixed and won't match the in-game collection
order.

Once a relic set is completed, the icon will be displayed in colour.

Before any part of a relic set has been acquired, a relic present box
icon is displayed.