Fix NoPaintingSkips always being set to enabled when connecting to AP
NoPaintingSkips used to have a 3rd, fully disabled state where neither
`painting_skips_on` or `painting_skips_off` were enabled. This extra
state was removed in 56053ee7f106 but the offset of `1` that was added
when reading the NoPaintingSkips value from slot data was not removed.

The offset has now been removed, so the NoPaintingSkips option is now
correctly set from slot data.