Fix out-of-logic act completion showing unlocked acts as in-logic
Completing an act out-of-logic would show any acts that were unlocked as
a result as being in-logic. This has been fixed by adding optional
requirements of logically being able to complete the prerequisite acts.
If the optional requirements are not met because the act was completed
out-of-logic, then the acts unlocked as a result will show as
out-of-logic.