Fix acts unlocked due to an out-of-logic free roam act showing as inaccessible
Free roam acts are always considered complete once the player has access
to the entrance. Having out-of-logic access to a free roam entrance
would not consider the free roam act complete, so any acts unlocked as
a result would show as inaccessible instead of out-of-logic.

The `completedActAt` logic function now returns an AccessibilityLevel
constant directly rather than true/false, this way, it can return the
AccessibilityLevel of the free roam entrance when the act at the
entrance being checked is a free roam act.