Ensure Tour is not used in logic when disabled
Chapters 6 and 7 get set to a Time Piece cost of 99999 when not
available, to effectively disable them, but Tour is accessible whenever
the Chapter 5 telescope is accessible, which is always an enabled
entrance, so the Time Rift - Tour entrance needs to be disabled more
strictly.

When Tour is disabled, it should be set to its vanilla entrance, so, by
disabling access to the entrance, this should disable access to Tour.