Add Free Roam act completions to the Entrances map
This makes it much easier to keep track of which Nyakuza Metro Free Roam
act completions still need to be completed to unlock the Chapter 7
Finale entrance.

Alpine Free Roam act completions are similar, but are easier to keep
track of due to there only being four.

Uses the diamond shape map locations to help indicate that the Free Roam
act completions are not randomizable entrances, so PopTracker version
0.26.2 or newer is recommended.