Only clear All Clear when the main objective has also been completed
It's not currently possible for this to happen because contracts are
only updated in datastorage when either the main objective is completed
or when the contract is fully cleared.

If it is changed in the future so that completing any part of a contract
updates datastorage, then adding in this extra check for the main
objective being completed will allow the tracker to continue working as
expected.