Show checkable, but unreachable, Time Rift entrances as blue
If the Time Rift entrance is unlocked, but inaccessible, and it is
possible to see in the Telescope what act is at the entrance, the
entrance now shows as blue.

For Blue Time Rifts, this is equivalent to saying that the Time Rift is
unlocked, but inaccessible.

For Purple Time Rifts, this is equivalent to saying that the Time Rift
is unlocked, but inaccessible, and that the Chapter Telescope is
accessible.

Alpine Skyline Blue Time Rifts never show as blue because being able to
unlock them implies being able to reach them.